+++ tests/Unit/EventSubscriber/UserCacheInvalidationSubscriberTest.php
@@ -27,6 +27,7 @@ use App\Entity\User;
 use App\EventSubscriber\UserCacheInvalidationSubscriber;
 use App\Exception\DoctrineResultCacheImplementationNotFoundException;
 use App\Generator\UserCacheIdGenerator;
+use App\Repository\UserRepository;
 use Doctrine\Common\Cache\Cache;
 use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
@@ -53,6 +54,9 @@ class UserCacheInvalidationSubscriberTest extends TestCase
     /** @var Cache|MockObject */
     private $resultCacheImplementation;
 
+    /** @var UserRepository|MockObject */
+    private $userRepository;
+
     protected function setUp(): void
     {
         parent::setUp();
@@ -61,6 +65,7 @@ class UserCacheInvalidationSubscriberTest extends TestCase
         $this->unitOfWork = $this->createMock(UnitOfWork::class);
         $this->userCacheIdGenerator = $this->createMock(UserCacheIdGenerator::class);
         $this->resultCacheImplementation = $this->createMock(Cache::class);
+        $this->userRepository = $this->createMock(UserRepository::class);
 
         $doctrineConfiguration = $this->createMock(Configuration::class);
         $doctrineConfiguration
@@ -75,7 +80,7 @@ class UserCacheInvalidationSubscriberTest extends TestCase
             ->method('getConfiguration')
             ->willReturn($doctrineConfiguration);
 
-        $this->subject = new UserCacheInvalidationSubscriber($this->userCacheIdGenerator);
+        $this->subject = new UserCacheInvalidationSubscriber($this->userCacheIdGenerator, $this->userRepository);
     }
 
     public function testSubscribedEvents(): void